Query Engine to  query distributed logs
I designed this project as a query engine to query and analyze
distributed logs from message based systems such as SCALA. The sample
logs are generated by SCALA actors from a separate project implementing
Gossip or PASTRY DHT. This is the netbeans project. I have also included
sample queries to give a sense of what sort of queries are currently
supported. But there are more features that I plan on adding.